apps/embeddings-extraction/requirements.txt
pytesseract
pillow
PyMuPDF
easyocr